models/ChartGemma/requirements.txt
# PyTorch with CUDA 12.1
torch==2.1.2+cu121
torchvision==0.16.2+cu121
torchaudio==2.1.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# Hugging Face stack
transformers==4.41.1
datasets==2.19.1

# LoRA, Quantization
peft==0.10.0
bitsandbytes==0.42.0

# PyTorch Lightning (PL)
pytorch-lightning==2.2.2
lightning==2.2.2

# Utilities
nltk==3.8.1
numpy==1.26.4
pillow==10.3.0
imageio==2.34.1

# For reproducibility
typing-extensions
